Remove duplicate and trivial tests, consolidate simple store tests
- Consolidate 8 trivial setter-only store tests into simple-stores.test.ts
- Remove 19 tautological default-value tests across all store files
- Remove duplicate clearQueueStore test and defaults from queue-storage
- Merge near-duplicate contextual tests (MiniPlayer, PlaylistScreen)
- Collapse 6 LibraryScreen getByTestId tests into 1
- Fold LoginFlow existence-check tests into structural renders
- Remove duplicate toggle-shuffle haptic test
- Merge null/undefined downloads tests into it.each
- Remove duplicate useRemoveFavorite haptic/onToggle tests
- Delete validation.test.ts (tested lodash, not app logic)
- Remove error-propagation tests that only test JS default behavior
- Merge null/undefined/ONLINE queue-utils tests into it.each
- Extract shared createTrack via mock-factories, add mock-storage setup
