+++ jest/functional/Player/playback.test.ts
@@ -234,45 +234,3 @@ describe('toggleRepeatMode', () => {
 		expect(TrackPlayer.pause).not.toHaveBeenCalled()
 	})
 })
-
-describe('togglePlayback error handling', () => {
-	beforeEach(() => {
-		jest.clearAllMocks()
-		;(usePlayerEngineStore as unknown as { getState: jest.Mock }).getState.mockReturnValue({
-			playerEngineData: PlayerEngine.REACT_NATIVE_TRACK_PLAYER,
-		})
-		// Also reset cast mock here to prevent leak from other describe blocks
-		const CastContext = require('react-native-google-cast').default
-		CastContext.getSessionManager.mockReturnValue({
-			getCurrentCastSession: jest.fn().mockResolvedValue(null),
-		})
-	})
-
-	it('handles TrackPlayer.getState rejection gracefully', async () => {
-		;(TrackPlayer.getState as jest.Mock).mockRejectedValue(new Error('Player not initialized'))
-
-		await expect(togglePlayback()).rejects.toThrow('Player not initialized')
-	})
-
-	it('handles TrackPlayer.play rejection gracefully', async () => {
-		;(TrackPlayer.getState as jest.Mock).mockResolvedValue({
-			currentState: 'paused',
-			totalDuration: 180,
-			currentPosition: 42,
-		})
-		;(TrackPlayer.play as jest.Mock).mockRejectedValue(new Error('Playback failed'))
-
-		await expect(togglePlayback()).rejects.toThrow('Playback failed')
-	})
-
-	it('handles TrackPlayer.pause rejection gracefully', async () => {
-		;(TrackPlayer.getState as jest.Mock).mockResolvedValue({
-			currentState: 'playing',
-			totalDuration: 180,
-			currentPosition: 42,
-		})
-		;(TrackPlayer.pause as jest.Mock).mockRejectedValue(new Error('Pause failed'))
-
-		await expect(togglePlayback()).rejects.toThrow('Pause failed')
-	})
-})
